feat: move core files to core folder
