+++ async-race/src/core/component/Component.ts
