+++ async-race/src/core/virtual-node/VirtualNode.ts
