+++ async-race/src/index.ts
@@ -1,13 +1,13 @@
 /* eslint-disable class-methods-use-this */
 /* eslint-disable import/extensions */
 
-import VirtualNode from './scripts/core/virtual-node/VirtualNode';
-import renderDOM from './scripts/core/render/renderDOM';
-import Component from './scripts/core/component/Component';
+import VirtualNode from './core/virtual-node/VirtualNode';
+import renderDOM from './core/render/renderDOM';
+import Component from './core/component/Component';
 import API from './scripts/API/Api';
 import CarE from './car';
 
-import { IState } from './scripts/core/component/types';
+import { IState } from './core/component/types';
 import { CarType } from './scripts/API/types';
 
 interface IAppState extends IState {
